requirements.txt
asgiref==3.8.1
Django==5.1.5
django-js-asset==3.0.1
django-mptt==0.16.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.1
